fix #963 Added import of CRUD transactions for SQL connector

+++ eclipse-plugin-studio/src/com/twinsoft/convertigo/eclipse/editors/connector/SqlConnectorDesignComposite.java
@@ -20,9 +20,13 @@
 package com.twinsoft.convertigo.eclipse.editors.connector;
 
 import java.io.File;
+import java.util.Arrays;
+import java.util.LinkedList;
+import java.util.List;
 
 import javax.swing.event.EventListenerList;
 
+import org.eclipse.jface.layout.RowLayoutFactory;
 import org.eclipse.swt.SWT;
 import org.eclipse.swt.events.KeyEvent;
 import org.eclipse.swt.events.MouseAdapter;
@@ -30,6 +34,7 @@ import org.eclipse.swt.events.MouseEvent;
 import org.eclipse.swt.graphics.Cursor;
 import org.eclipse.swt.layout.GridData;
 import org.eclipse.swt.layout.GridLayout;
+import org.eclipse.swt.layout.RowLayout;
 import org.eclipse.swt.widgets.Button;
 import org.eclipse.swt.widgets.Composite;
 import org.eclipse.swt.widgets.Display;
@@ -51,6 +56,7 @@ import com.twinsoft.convertigo.eclipse.ConvertigoPlugin;
 import com.twinsoft.convertigo.eclipse.editors.CompositeEvent;
 import com.twinsoft.convertigo.eclipse.editors.CompositeListener;
 import com.twinsoft.convertigo.eclipse.views.projectexplorer.ProjectExplorerView;
+import com.twinsoft.convertigo.eclipse.views.projectexplorer.TreeObjectEvent;
 
 class SqlConnectorDesignComposite extends Composite {
 
@@ -59,67 +65,93 @@ class SqlConnectorDesignComposite extends Composite {
 	private TableColumn tblclmnCallableName;
 	private TableColumn tblclmnDescription;
 	private TableColumn tblclmnGroupName;
-	
+
 	private Button btnImportAsTransactions;
 
 	private SqlConnector sqlConnector;
-	
+
 	private ProjectExplorerView projectExplorerView = null;
-	
+
 	public SqlConnectorDesignComposite(Connector connector, Composite parent, int style) {
 		super(parent, style);
-		
+
 		sqlConnector = (SqlConnector)connector;
-		
+
 		// add ProjectExplorerView to the listeners of the associated sql connector
 		projectExplorerView = ConvertigoPlugin.getDefault().getProjectExplorerView();
 		if (projectExplorerView != null) {
 			addCompositeListener(projectExplorerView);
 		}
-		
+
 		GridLayout gridLayout = new GridLayout();
 		gridLayout.numColumns = 1;
 		this.setLayout(gridLayout);
-		
+
 		Composite composite = new Composite(this, SWT.NONE);
 		composite.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true, 1, 1));
 		composite.setLayout(new GridLayout(3, false));
-		
+
 		Label lblNewLabel = new Label(composite, SWT.NONE);
 		lblNewLabel.setToolTipText("Type Here a Stored Procedure/Function Pattern such as SUB%");
 		lblNewLabel.setLayoutData(new GridData(SWT.LEFT, SWT.CENTER, false, false, 1, 1));
 		lblNewLabel.setText("Search Pattern");
-		
+
 		text = new Text(composite, SWT.BORDER);
 		text.setText("%");
 		text.setLayoutData(new GridData(SWT.FILL, SWT.CENTER, true, false, 1, 1));
+
+		Button btnNewButton = new Button(composite, SWT.NONE);
+		btnNewButton.setText("Search");
+
+		lblNewLabel = new Label(composite, SWT.NONE);
+		lblNewLabel.setLayoutData(new GridData(SWT.LEFT, SWT.CENTER, false, false, 1, 1));
+		lblNewLabel.setText("Search in…");
+
+		var types = new Composite(composite, SWT.NONE);
+		types.setLayoutData(new GridData(SWT.LEFT, SWT.CENTER, false, false, 2, 1));
+		types.setLayout(new RowLayout());
+
+		var checkbox = new Button(types, SWT.CHECK);
+		checkbox.setText("TABLE");
+		checkbox.setSelection(true);
+		checkbox = new Button(types, SWT.CHECK);
+		checkbox.setText("PROCEDURE");
+		checkbox.setSelection(true);
+		checkbox = new Button(types, SWT.CHECK);
+		checkbox.setText("FUNCTION");
+		checkbox.setSelection(true);
+
+
 		text.addKeyListener(new org.eclipse.swt.events.KeyAdapter() {
 			@Override
 			public void keyReleased(KeyEvent e) {
 				if (e.character == SWT.CR) {
-					search(text.getText());
+					btnNewButton.notifyListeners(SWT.MouseDown, null);
 				}
 			}
 		});
-		
-		Button btnNewButton = new Button(composite, SWT.NONE);
-		btnNewButton.setText("Search");
 		btnNewButton.addMouseListener(new MouseAdapter() {
 			@Override
 			public void mouseDown(MouseEvent e) {
-				search(text.getText());
+				var enabled = new LinkedList<String>();
+				for (var c: types.getChildren()) {
+					if (c instanceof Button b && b.getSelection()) {
+						enabled.add(b.getText());
+					}
+				}
+				search(text.getText(), enabled);
 			}
 		});
-		
+
 		table = new Table(composite, SWT.BORDER | SWT.FULL_SELECTION | SWT.MULTI);
 		table.setLayoutData(new GridData(SWT.FILL, SWT.FILL, true, true, 3, 1));
 		table.setHeaderVisible(true);
 		table.setLinesVisible(true);
-		
+
 		tblclmnCallableName = new TableColumn(table, SWT.NONE);
 		tblclmnCallableName.setWidth(200);
 		tblclmnCallableName.setText("Name");
-		
+
 		tblclmnDescription = new TableColumn(table, SWT.NONE);
 		tblclmnDescription.setWidth(500);
 		tblclmnDescription.setText("Description");
@@ -127,14 +159,51 @@ class SqlConnectorDesignComposite extends Composite {
 		tblclmnGroupName = new TableColumn(table, SWT.NONE);
 		tblclmnGroupName.setWidth(200);
 		tblclmnGroupName.setText("Type");
+
+		var footer = new Composite(composite, SWT.NONE);
+		footer.setLayoutData(new GridData(SWT.FILL, SWT.CENTER, true, false, 3, 1));
+		footer.setLayout(new GridLayout(3, false));
+
+		lblNewLabel = new Label(footer, SWT.NONE);
+		lblNewLabel.setLayoutData(new GridData(SWT.LEFT, SWT.CENTER, false, false, 1, 1));
+		lblNewLabel.setText("For TABLE…");
+
+		var cruds = new Composite(footer, SWT.NONE);
+		cruds.setLayoutData(new GridData(SWT.LEFT, SWT.CENTER, false, false, 1, 1));
+		cruds.setLayout(new RowLayout());
+
+		checkbox = new Button(cruds, SWT.CHECK);
+		checkbox.setText("INSERT");
+		checkbox.setSelection(true);
+		checkbox = new Button(cruds, SWT.CHECK);
+		checkbox.setText("SELECT");
+		checkbox.setSelection(true);
+		checkbox = new Button(cruds, SWT.CHECK);
+		checkbox.setText("UPDATE");
+		checkbox.setSelection(true);
+		checkbox = new Button(cruds, SWT.CHECK);
+		checkbox.setText("DELETE");
+		checkbox.setSelection(true);
 		
-		btnImportAsTransactions = new Button(composite, SWT.NONE);
+		var right = new Composite(footer, SWT.NONE);
+		right.setLayoutData(new GridData(SWT.RIGHT, SWT.CENTER, true, false, 1, 1));
+		right.setLayout(RowLayoutFactory.fillDefaults().center(true).create());
+		var override = new Button(right, SWT.CHECK);
+		override.setText("Override");
+		
+		btnImportAsTransactions = new Button(right, SWT.NONE);
 		btnImportAsTransactions.setText("Import as transaction(s) in project");
-		btnImportAsTransactions.setLayoutData(new GridData(SWT.LEFT, SWT.CENTER, false, false, 3, 1));
+
 		btnImportAsTransactions.addMouseListener(new MouseAdapter() {
 			@Override
 			public void mouseDown(MouseEvent e) {
-				createSqlTransactions(table.getSelection());
+				var enabled = new LinkedList<String>();
+				for (var c: cruds.getChildren()) {
+					if (c instanceof Button b && b.getSelection()) {
+						enabled.add(b.getText());
+					}
+				}
+				createSqlTransactions(table.getSelection(), override.getSelection(), enabled);
 			}
 		});
 	}
@@ -157,44 +226,44 @@ class SqlConnectorDesignComposite extends Composite {
 		}
 	}
 
-	private void search(String pattern) {
+	private void search(String pattern, List<String> types) {
 		Display display = Display.getDefault();
-		Cursor waitCursor = new Cursor(display, SWT.CURSOR_WAIT);		
+		Cursor waitCursor = new Cursor(display, SWT.CURSOR_WAIT);
 		Shell shell = display.getActiveShell();
 		if (shell != null) {
 			try {
 				shell.setCursor(waitCursor);
-				
-			    ConvertigoPlugin.logDebug("Searching repository...");
-				Document doc = SqlConnector.executeSearch(sqlConnector, pattern);
-		        ConvertigoPlugin.logDebug("Search done.");
-		        
-		        if (doc != null) {
-		        	table.removeAll();
-		        	
-		        	NodeList items = doc.getDocumentElement().getElementsByTagName("item");
-		        	for (int i=0; i<items.getLength(); i++) {
-		        		Element item = (Element) items.item(i);
-		        		
-		        		Element func = (Element) item.getElementsByTagName("NAME").item(0);
-		        	    String funcName = func.getTextContent();
-		        	    String specific_name = func.getAttribute("specific_name");
-		        	    
-		        	    String groupName ="";
-		        	    try {
-		        	    	groupName = item.getElementsByTagName("TYPE").item(0).getTextContent();
-		        	    } catch (Exception e) {}
-		        	    String sText="";
-		        	    try {
-		        	    	sText = item.getElementsByTagName("REMARKS").item(0).getTextContent();
-		        	    } catch (Exception e) {}
-		        	    
-	      				TableItem tableItem = new TableItem(this.table, SWT.NONE);
-	    			    tableItem.setText(new String[] {funcName,sText,groupName});
-	    			    tableItem.setData("specific_name", specific_name);
-		        	}
-		        }
-		        
+
+				ConvertigoPlugin.logDebug("Searching repository...");
+				Document doc = SqlConnector.executeSearch(sqlConnector, pattern, types);
+				ConvertigoPlugin.logDebug("Search done.");
+
+				if (doc != null) {
+					table.removeAll();
+
+					NodeList items = doc.getDocumentElement().getElementsByTagName("item");
+					for (int i=0; i<items.getLength(); i++) {
+						Element item = (Element) items.item(i);
+
+						Element func = (Element) item.getElementsByTagName("NAME").item(0);
+						String funcName = func.getTextContent();
+						String specific_name = func.getAttribute("specific_name");
+
+						String groupName ="";
+						try {
+							groupName = item.getElementsByTagName("TYPE").item(0).getTextContent();
+						} catch (Exception e) {}
+						String sText="";
+						try {
+							sText = item.getElementsByTagName("REMARKS").item(0).getTextContent();
+						} catch (Exception e) {}
+
+						TableItem tableItem = new TableItem(this.table, SWT.NONE);
+						tableItem.setText(new String[] {funcName,sText,groupName});
+						tableItem.setData("specific_name", specific_name);
+					}
+				}
+
 			} catch (Exception ee) {
 				ConvertigoPlugin.logException(ee, "Error while searching repository");
 			} finally {
@@ -203,10 +272,10 @@ class SqlConnectorDesignComposite extends Composite {
 			}
 		}
 	}
-	
-	private void createSqlTransactions(final TableItem[] items) {
+
+	private void createSqlTransactions(final TableItem[] items, boolean override, List<String> cruds) {
 		Display display = Display.getDefault();
-		Cursor waitCursor = new Cursor(display, SWT.CURSOR_WAIT);		
+		Cursor waitCursor = new Cursor(display, SWT.CURSOR_WAIT);
 		Shell shell = display.getActiveShell();
 		if (shell != null) {
 			try {
@@ -214,31 +283,44 @@ class SqlConnectorDesignComposite extends Composite {
 				for (int i=0; i < items.length; i++) {
 					TableItem item = items[i];
 					String callableName = item.getText(0);
-					String callableDesc = item.getText(1);
-					String specific_name = (String) item.getData("specific_name");
-					ConvertigoPlugin.logDebug("Creating transaction for CALL '"+callableName+"' ...");
-					
-					if (specific_name.isEmpty()) {
-						specific_name = callableName;
-					}
-					
-					SqlTransaction sqlTransaction = SqlConnector.createSqlTransaction(sqlConnector, callableName, specific_name);
-					if (sqlTransaction != null) {
-						Transaction transaction = sqlConnector.getTransactionByName(sqlTransaction.getName());
-						if (transaction != null) {
-							try {
-								File xsdFile = new File(transaction.getSchemaFilePath());
-								if (xsdFile.exists()) {
-									xsdFile.delete();
-								}
-							}
-							catch (Exception e) {}
-							sqlConnector.remove(transaction);			
+					String type = item.getText(2);
+					List<SqlTransaction> sqlTransactions = null;
+					if ("TABLE".equals(type)) {
+						sqlTransactions = SqlConnector.createSqlTransaction(sqlConnector, callableName, cruds);
+					} else {
+						String specific_name = (String) item.getData("specific_name");
+						ConvertigoPlugin.logDebug("Creating transaction for CALL '"+callableName+"' ...");
+
+						if (specific_name.isEmpty()) {
+							specific_name = callableName;
 						}
-						sqlTransaction.setComment(callableDesc);
-						sqlConnector.add(sqlTransaction);
-						fireObjectChanged(new CompositeEvent(sqlConnector));
-						ConvertigoPlugin.logDebug("Transaction added.");
+						sqlTransactions = Arrays.asList(SqlConnector.createSqlTransaction(sqlConnector, callableName, specific_name));
+					}
+					for (var sqlTransaction: sqlTransactions) {
+						if (sqlTransaction != null) {
+							Transaction transaction = sqlConnector.getTransactionByName(sqlTransaction.getName());
+							if (override || transaction == null) {
+								if (transaction != null) {
+									try {
+										File xsdFile = new File(transaction.getSchemaFilePath());
+										if (xsdFile.exists()) {
+											xsdFile.delete();
+										}
+									}
+									catch (Exception e) {}
+									sqlConnector.remove(transaction);
+								}
+								sqlConnector.add(sqlTransaction);
+								fireObjectChanged(new CompositeEvent(sqlConnector));
+								var pev = ConvertigoPlugin.getDefault().getProjectExplorerView();
+								if (pev != null) {
+									var dbot = pev.findTreeObjectByUserObject(sqlTransaction);
+									pev.fireTreeObjectPropertyChanged(new TreeObjectEvent(dbot, "sqlQuery", "", sqlTransaction.getSqlQuery(), TreeObjectEvent.UPDATE_NONE));
+									pev.setSelectedTreeObject(dbot);
+								}
+								ConvertigoPlugin.logDebug("Transaction added.");
+							}
+						}
 					}
 				}
 			} catch (Exception ee) {
